Add ability to pass properties to `ApiClient`.
This allows us to pass config options to the `ApiClient`.
Currently this is used to allow us to tweak the polling
interval for the data exports so the integration specs
don't take forever.  In the future, this needs to be cleaned
up to make the config a first class citizen.